fix(topography): coerce NetCDF lmax/radius attribute types
load_topography_from_file read lmax via ds.attrib["lmax"], which NCDatasets
returns as Int32, but TopographyField only has an Int (Int64) constructor ->
MethodError loading any topo file with an lmax attribute. Same latent issue
for a Float32 radius attribute. Wrap both in Int(...)/Float64(...). Regression
test in test/topography_load_attr.jl (surfaced by the coverage sweep).

+++ src/bcs/topography/topography_data.jl
@@ -449,12 +449,14 @@ function load_topography_from_file(filename::String, location::BoundaryLocation;
     end
 
     NCDataset(filename, "r") do ds
-        # Read lmax
-        lmax = haskey(ds.attrib, "lmax") ? ds.attrib["lmax"] : 32
+        # Read lmax. NetCDF stores integer attributes as 32-bit, so this comes
+        # back as Int32; TopographyField's ctor requires Int — convert explicitly.
+        lmax = haskey(ds.attrib, "lmax") ? Int(ds.attrib["lmax"]) : 32
 
-        # Read radius
+        # Read radius. Likewise coerce the (possibly Float32) attribute to Float64
+        # to match the TopographyField constructor.
         r = radius !== nothing ? radius :
-            (haskey(ds.attrib, "radius") ? ds.attrib["radius"] : 1.0)
+            (haskey(ds.attrib, "radius") ? Float64(ds.attrib["radius"]) : 1.0)
 
         field = TopographyField(lmax, lmax, r, location)
